kgsl: Add support to compile kgsl with upstream kernel
A lot of functions, macros and header files are used by kgsl that are only
present in the downstream qcom kernel. Add a new include directory that
contains the downstream header files with stubbed functions and
macros required for compilation. The header files are modified versions
of the downstream header files that only contain the functions and
macros used by kgsl. This new directory can be included in the conf file
used to compile kgsl with upstream kernel.

Change-Id: I071248e074cdcbdc5f586a47b262bd585d7fa596
Signed-off-by: Puranam V G Tejaswi <quic_pvgtejas@quicinc.com>