kgsl: Add a device link between cx power domain and iommu device
As part of runtime_resume PM callback, the smmu driver resets the device.
This creates problems if GPU is active and runtime_resume callback of
iommu device kicks in. Hence, add device link between the CX power domain
and iommu device to keep the iommu device active when GPU is active. This
is not needed when using downstream driver with qcom addons because reset
of the device is done only during resume PM callback.

Change-Id: Ib2b725a4cfe675e8440de5fafa92935f9a2fcdba
Signed-off-by: Puranam V G Tejaswi <quic_pvgtejas@quicinc.com>