kgsl: hwsched: Add hw fence shadow table support
GMU FW needs to allocate a shadow table to cache external
fences that it receives in its rx queue. It will send a
mem_alloc request of type MEMKIND_HW_FENCE_SHADOW. The number
of entries in this table is derived from a hw-fence dt property
and communicated to GMU via VRB. Additionally, the hash index
of an input fence must be communicated to GMU via the
hfi_syncobj packet. This allows for constant time look up when
checking whether an external input fence is signaled or not in
the GMU FW. Hardware fence feature requires that GMU has the
shadow table support. Hence, check for the GMU FW version when
enabling this feature.

Change-Id: Ic63500481c484941c4f02942bfe5b2226690c9a4
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>