kgsl: hwfence: Take detached context refcount
When we encounter a failure during context detach, we move the
pending hardware fences from this context to a list. After
reset is complete, we trigger these fences one by one. As part of
freeing these hardware fences, we decrement the drawctxt's
hw_fence count. This can cause a use-after-free of the drawctxt
structure since this context is already detached and may have
been freed. To fix this, take the context refcount for each of
these hardware fences in the detach path, and put them back when
each of these hardware fences are triggered after recovery.

Change-Id: I3077be6b07cc55c77329c06330433d9ece9569c8
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>
Signed-off-by: Vaishali Gupta <quic_vai@quicinc.com>