kgsl: hwfence: Fix indexing issue when handling F2H_MSG_SYNCOBJ_QUERY
As part of handling F2H_MSG_SYNCOBJ_QUERY packet in kgsl, we currently
are not ignoring signaled sw-fences that might be part of the queried
sync object. This can be fixed by skipping any software dma fences
in the sync object as part of handling F2H_MSG_SYNCOBJ_QUERY packet.
However, a better fix is to keep one-to-one mapping between each
hardware fence and its index when composing the H2F_MSG_ISSUE_SYNCOBJ
packet. We can easily achieve this by introducing an array of hardware
fences for each sync object. This also simplifies how we walk a sync
object to find all the hardware fences in it.
Define new macros to derive the maximum number of hardware fences that
can be embedded in a single H2F_MSG_ISSUE_SYNCOBJ packet. If a sync
object contains more than the capped number of hardware fences then
set KGSL_SYNCOBJ_SW flag for this sync object.

Change-Id: I4ab03fab82475e5d35ebb785b41a19265661aa45
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>