kgsl: Poll CX CFG GDSC register to ensure CX GDSC collapse
Currently, KGSL polls for the PWR_ON bit transition from 1 to 0
to ensure CX GDSC collapse. However, this bit only indicates that
the power-down sequence has begun and doesn't guarantee that the CX
collapsed at the hardware level. To address this, we need to poll for
the POWER_DOWN_COMPLETE bit of the CX CFG GDSC register. This ensures
that the CX collapses properly at the hardware level. Thus, update the
CX GDSC polling logic accordingly.

Change-Id: I6cee228e50e6c9ca051163137e2c00399ee9f3b9
Signed-off-by: Sanjay Yadav <quic_sanjyada@quicinc.com>