kgsl: hwfence: Get kgsl fence context refcount
Some sync objects can contain input fence which is a kgsl fence and
belongs to a kgsl context. It is possible that this context gets
detached and destroyed before GMU processes this sync object.
This can create functional issues since the memstore eop/sop of
this context will not be reliable anymore. Hence, take a refcount
of this context before dispatching this sync object to GMU, and put
it back only when we know that GMU has retired this sync object. If
the refcount fails, then it means that the context is going away
and we can deem this fence as signaled in that case.

Change-Id: I70235873caa1c0f3f46df39b1aceee9c9e41c168
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>