kgsl: iommu: Use map_pages() callback correctly
Currently, when using map_pages() callback in _iopgtbl_map_sg(), the
entire size equivalent to sg->length is being passed. It is possible
that success is returned but the entire size is not mapped. For
instance, when size of 2 MB is passed, only 1 MB is mapped and success
is returned. Hence, use map_pages() in a loop until the entire size is
mapped.

In _iopgtbl_map_page_to_range(), map_size is being passed for each
iteration without being initialized to zero every time. This results in
incorrect calculation of the overall mapped memory. Fix this by passing
mapped variable directly to the callback. Use similar approach for
_iopgtbl_map_pages() and remove redundant variable size.

Change-Id: I1eb802e06b8c0364d4956ea27efc3cf67bb67a7d
Signed-off-by: Puranam V G Tejaswi <quic_pvgtejas@quicinc.com>