kgsl: hwfence: Avoid scenarios that call kgsl_hw_fence_destroy()
Currently, we create a hardware fence first and then destroy it in
some scenarios for example if the timestamp is already retired, or
if there isn't enough space in context's hw fence buffer. This
causes a hardware fence leak because a hardware fence which is never
sent to tx queue is never deleted from the global table. To fix this,
call kgsl_hw_fence_create() where we are sure that the hardware
fence will be sent to GMU and eventually be sent to tx queue.

Change-Id: I05677a3441e742f7b4d93a6c23a3800ecf07dd78
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>