kgsl: gen7: Fix GEN7 register programming
Currently, the BV pipe of the GEN7_GRAS_NC_MODE_CNTL
register is not programmed correctly. Due to this it
shows reset value. To fix this, program both the BR and BV
pipes using the GEN7_CP_APERTURE_CNTL_HOST then, write to
the GEN7_GRAS_NC_MODE_CNTL register for BR and BV respectively.

For gen7 there are pipe specific registers which need
to be part of the static powerup list for CP to restore
at IFPC interval. Hence add support to include all the
pipe register to external powerup reglist.

Change-Id: I756b0fb38cf4b808203ee4bc5a00c08884ce2c4e
Signed-off-by: Sanjay Yadav <quic_sanjyada@quicinc.com>