kgsl: Reseed GPU CX timer after suspend resume
When the device enters in suspend state, the CX can be collapsed causing
the GPU CX timer to pause. To ensure synchronization with the CPU timer,
reseed the GPU CX timer after resuming from suspend or hibernation. This
ensures that the GPU CX timer remains in sync with the CPU timer.

Change-Id: Idda014c1878e22eebc4c500fbc9d836fe6d068a4
Signed-off-by: Pankaj Gupta <quic_gpankaj@quicinc.com>