msm: kgsl: Fix few VBO issues
If sharedmem cannot be unmapped from the mmu, it can still be accessed
by the GPU. Therefore it is not safe to free the backing memory. In the
case that unmap fails, do not free it or return it to the system. If the
VBO range cannot be unmapped the underlying physical memory should not
be freed. Enforce this by not decrementing the mem entry's refcount.

Change-Id: Id466152cf7b318ad254004e496a4afafc5be5576
Signed-off-by: NISARG SHETH <quic_nsheth@quicinc.com>