kgsl: Make fault recovery and suspend mutually exclusive
It is possible that kgsl is doing fault recovery when a system suspend
request comes in. The system suspend path increments power domain
prepared_count and that causes fault recovery to fail because the cx
and gx gdscs are not successfully toggled. To fix this, introduce a
new mutex to make sure fault recovery and system suspend are mutually
exclusive. Also, introduce a fault recovery completion to make sure
we don't attempt fault recovery until we are sure that the system path
will not be able to increment the power domain prepared_count while
fault recovery is in place.

Change-Id: I0eb51e7afb3147d195ba4b19bcb0f2087ed6d752
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>