kgsl: hwfence: Fix out of bound access during context queue read
The context queue in kgsl is a circular buffer. However, the `queue`
variable in `find_timeout_syncobj()` and the `cmd` variable in
`log_syncobj()` are decoded without considering the circular nature
of the context queue. As a result, this can lead to out-of-bounds
access. Fix it by introducing a function that correctly reads the
context queue.

Change-Id: I751f3b16782b56a5a32efe616d98f1e2fe4dd781
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>