kgsl: pwrctrl: Add sub_type constraint check to update expiry
When a context is submitted with a MAX constraint sub_type, followed by
back-to-back submissions of a MIN constraint sub_type, the expiry of the
MAX constraint is increased. This results in the max_pwrlevel being
retained for a longer period.
To fix this, add sub_type constraint check also. This ensures that
expiry is updated only if the previous and current sub_type constraint
are the same.

Change-Id: I46cd6efa3c2bd698bf26febfb1928f4fcd8ce30b
Signed-off-by: Sanjay Yadav <quic_sanjyada@quicinc.com>