kgsl: pwrctrl: Use dedicated GMU CX GenPD instance for voting
Currently, there is a race condition in GenPD framework where
GPU CX GDSC can remain ON if both GMU and KGSL SMMU devices are
suspending in parallel and are voting on the same power domain.
Use dedicated GMU CX GenPD instance for CX GDSC voting to avoid
the race. Since GMU and SMMU votes are consolidate in CX GenPD
instance, use it for registering the notifier.

Change-Id: I220dfd94bb7d5155fc4989047083c70b3560f3c2
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>