kgsl: Handle CP halt in GPU recovery
As part of GPU recovery sequence, CP is not halted for all
faults in sw scheduler path. This can lead to parallel
execution of instructions in CP while GPU recovery sequence
is in progress. So, halt CP for all faults. In hw scheduler
path, halt CP for page faults. CP is halted from GMU when
required for other faults.

Also, there is a corner case wherein, if CP is halted while
executing SMMU_TABLE_UPDATE packet, there is a chance of recovery
failure. The problematic window is right after the page table walk
disable and before it is enabled in SMMU_TABLE_UPDATE packet.
So, enable page table walk in user and lpac contexts during boot.

Change-Id: Id8a63a9518312d2b314acde244327e034bed144f
Signed-off-by: Raviteja Narayanam <quic_ravnar@quicinc.com>