kgsl: hwsched: Fix pagefault recovery in SMMU TERMINATE model
In TERMINATE model, we are currently relying on the smmu fault handler
to clear FSR to resume SMMU transactions. Which means if smmu fault handler
takes a while to finish, the SMMU will keep terminating transactions. Say
at the same time kgsl attempts recovery, then recovery would fail during
CP INIT as the GPU SMMU is still terminating transactions. To address this,
first make sure the smmu handler doesn't clear FSR in the TERMINATE model.
Secondly, add means to clear the FSR in kgsl, before attempting CP INIT.
Also make sure that we don't run crashdumper if we detect that GC
SMMU is in TERMINATE mode in presence of an outstanding pagefault.

Change-Id: I8b7a96492760ea4114dfdc73ef2b6d6228b36e7d
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>
Signed-off-by: Pankaj Gupta <quic_gpankaj@quicinc.com>