kgsl: Fix length check for qcom,ipc-core property
Currently, the length check for the qcom,ipc-core property is
incorrect. The length is being compared directly to the sum of
addr_cells and size_cells, which are in terms of 32-bit cells.
This results in a mismatch. Therefore, correct the length check
by multiplying the sum of addr_cells and size_cells by sizeof(__be32)
to ensure the length is compared correctly in terms of bytes.

Change-Id: I091004723d7a338f9749b34e424d7895da0e2d94
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>
Signed-off-by: Pankaj Gupta <quic_gpankaj@quicinc.com>