kgsl: gmu: Handle dummy gMxC resource entries from cmd_db
For targets where gMxC is not supported but RM still has a hardware
instance, AOP will update this as a dummy entry in cmd_db. This means
cmd_db will have an entry for gMxC with all-zero values (no supported
operating levels). To handle this in KGSL, add a check for dummy gMxC
resource entries from cmd_db. If such an entry is found, return an
error code (-ENODATA) instead of attempting to process it, which would
lead to an invalid MxC vote.

Change-Id: I2efa3924bf7e0da1edf615ef0523ced883ad98b5
Signed-off-by: SIVA MULLATI <quic_smullati@quicinc.com>