kgsl: Fix imported memory calculation for shared buffers
Currently, when the same buffer is mapped as a surface in one
process, it is ignored in the imported memory calculation of
another process that imports it as an image, causing discrepancies.

To fix this, the memory object size is now divided by the sum of EGL
surface and EGL image counts. This ensures that the imported memory is
accurately accounted for across different processes, regardless of flags.

Change-Id: Ieade58964c51758f7fe54b2a473aa03fe374482e
Signed-off-by: Sanjay Yadav <quic_sanjyada@quicinc.com>