kgsl: hwsched: Dump GMU context queue on unregister timeout
Currently, GMU context queue is not dumped in snapshot for
context whose context unregister with GMU fails. KGSL dumps
the GMU context queue for contexts that are registered with
GMU. If the context unregister fails, we set the flag to false,
and thus it is not dumped. Fix this by updating the flag only
when the context unregister is successful. During GPU reset for
context unregister timeout, all contexts will be unregistered.

Change-Id: Ic896553f18395df5903ae42725cf02735dded73a
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>