kgsl: Ensure task policy and priority are restored correctly
During the GPU snapshot process, there is no need for the task to
run at RT priority as it might hog the system. Currently, the task's
priority is downgraded to normal and then bumped back to priority 50
at the end of the snapshot. However, the task's original scheduling
attributes are not being restored to their original state, which could
lead to incorrect scheduling behavior if the original task was using a
different policy or priority.

This fix ensures that both the scheduling policy and priority of the
current task are correctly saved and restored during the GPU snapshot
process.

Change-Id: I58f5b706f9d0532423cd034f959d0da12030032f
Signed-off-by: Pankaj Gupta <quic_gpankaj@quicinc.com>