kgsl: hwsched: Reset the return code when polling for the start ack
When polling for the start ack, we may receive a F2H_MSG_PROCESS_TRACE
packet. Current code does not reset the stale return code in this case, so
we exit the polling loop in error. Reset the return code so that we
continue polling till we need to correctly exit the polling loop.

Change-Id: Id5399c93dd20de539fccb232ad4644e6f8003ad9
Signed-off-by: Lynus Vaz <quic_lvaz@quicinc.com>